fix(consensus): fix pr
