+++ chain/deputynode/manager.go
@@ -22,6 +22,7 @@ var (
 	ErrQueryFutureTerm       = errors.New("can't query future term")
 	ErrMineGenesis           = errors.New("can not mine genesis block")
 	ErrNotDeputy             = errors.New("not a deputy address in specific height")
+	ErrSmallerMineTime       = errors.New("the time of block must not be smaller than parent's")
 )
 
 type BlockLoader interface {
@@ -245,4 +246,92 @@ func (m *Manager) GetDeputyByDistance(targetHeight uint32, parentBlockMiner comm
 		}
 	}
 	return nil, ErrNotDeputy
+}
+
+// GetMinerDistance get miner index distance. It is always greater than 0 and not greater than deputy count
+func (m *Manager) GetMinerDistance(targetHeight uint32, parentBlockMiner, targetMiner common.Address) (uint32, error) {
+	if targetHeight == 0 {
+		return 0, ErrMineGenesis
+	}
+	deputies := m.GetDeputiesByHeight(targetHeight)
+	nodeCount := uint32(len(deputies))
+
+	// find target block miner deputy
+	targetDeputy := findDeputyByAddress(deputies, targetMiner)
+	if targetDeputy == nil {
+		return 0, ErrNotDeputy
+	}
+
+	// Genesis block is pre-set, not belong to any deputy node. So only blocks start with height 1 is mined by deputies
+	// The reward block changes deputy nodes, so we need recompute the slot
+	if targetHeight == 1 || IsRewardBlock(targetHeight) {
+		return targetDeputy.Rank + 1, nil
+	}
+
+	// if they are same miner, then return deputy count
+	if targetMiner == parentBlockMiner {
+		return nodeCount, nil
+	}
+
+	// find last block miner deputy
+	lastDeputy := findDeputyByAddress(deputies, parentBlockMiner)
+	if lastDeputy == nil {
+		return 0, ErrNotDeputy
+	}
+	return (nodeCount + targetDeputy.Rank - lastDeputy.Rank) % nodeCount, nil
+}
+
+func findDeputyByAddress(deputies []*types.DeputyNode, addr common.Address) *types.DeputyNode {
+	for _, node := range deputies {
+		if node.MinerAddress == addr {
+			return node
+		}
+	}
+	return nil
+}
+
+// GetNextMineWindow get next time window to mine block. The times are timestamps in millisecond
+func (m *Manager) GetNextMineWindow(nextHeight uint32, distance uint32, parentTime int64, currentTime int64, mineTimeout int64) (int64, int64) {
+	nodeCount := m.GetDeputiesCount(nextHeight)
+	// 所有节点都超时所需要消耗的时间，也可以看作是下一轮出块的开始时间
+	oneLoopTime := int64(nodeCount) * mineTimeout
+	// 网络传输耗时，即当前时间减去父块区块头中的时间戳
+	passTime := currentTime - parentTime
+	if passTime < 0 {
+		passTime = 0
+	}
+	// 从父块开始，经过的整轮数
+	passLoop := passTime / oneLoopTime
+	// 可以出块的时间窗口
+	windowFrom := parentTime + passLoop*oneLoopTime + int64(distance-1)*mineTimeout
+	windowTo := parentTime + passLoop*oneLoopTime + int64(distance)*mineTimeout
+	if windowTo <= currentTime {
+		windowFrom += oneLoopTime
+		windowTo += oneLoopTime
+	}
+
+	log.Debug("GetNextMineWindow", "windowFrom", windowFrom, "windowTo", windowTo, "parentTime", parentTime, "passTime", passTime, "distance", distance, "passLoop", passLoop, "nodeCount", nodeCount)
+	return windowFrom, windowTo
+}
+
+// GetCorrectMiner get the correct miner to mine a block after parent block
+func (m *Manager) GetCorrectMiner(parent *types.Header, mineTime int64, mineTimeout int64) (common.Address, error) {
+	if mineTime < 1e10 {
+		panic("mineTime should be milliseconds")
+	}
+	passTime := mineTime - int64(parent.Time)*1000
+	if passTime < 0 {
+		return common.Address{}, ErrSmallerMineTime
+	}
+	nodeCount := m.GetDeputiesCount(parent.Height + 1)
+	// 所有节点都超时所需要消耗的时间，也可以看作是下一轮出块的开始时间
+	oneLoopTime := int64(nodeCount) * mineTimeout
+	minerDistance := (passTime%oneLoopTime)/mineTimeout + 1
+
+	deputy, err := m.GetDeputyByDistance(parent.Height+1, parent.MinerAddress, uint32(minerDistance))
+	if err != nil {
+		return common.Address{}, err
+	}
+	log.Debug("GetCorrectMiner", "correctMiner", deputy.MinerAddress, "parent", parent.MinerAddress, "mineTime", mineTime, "mineTimeout", mineTimeout, "passTime", passTime, "nodeCount", nodeCount)
+	return deputy.MinerAddress, nil
 }
