+++ chain/miner/miner.go
@@ -1,7 +1,6 @@
 package miner
 
 import (
-	"github.com/LemoFoundationLtd/lemochain-core/chain/consensus"
 	"github.com/LemoFoundationLtd/lemochain-core/chain/deputynode"
 	"github.com/LemoFoundationLtd/lemochain-core/chain/types"
 	"github.com/LemoFoundationLtd/lemochain-core/common"
@@ -53,13 +52,13 @@ func New(cfg MineConfig, chain Chain, dm *deputynode.Manager, txPool TxPool) *Mi
 		blockInterval:           cfg.SleepTime,
 		timeoutTime:             cfg.Timeout,
 		reservedPropagationTime: cfg.ReservedPropagationTime,
-		chain:          chain,
-		dm:             dm,
-		txPool:         txPool,
-		recvNewBlockCh: make(chan *types.Block, 1),
-		timeToMineCh:   make(chan *MineInfo),
-		stopCh:         make(chan struct{}),
-		quitCh:         make(chan struct{}),
+		chain:                   chain,
+		dm:                      dm,
+		txPool:                  txPool,
+		recvNewBlockCh:          make(chan *types.Block, 1),
+		timeToMineCh:            make(chan *MineInfo),
+		stopCh:                  make(chan struct{}),
+		quitCh:                  make(chan struct{}),
 	}
 }
 
@@ -183,7 +182,7 @@ func (m *Miner) getSleepTime(mineHeight uint32, distance uint32, parentTime int6
 	// 网络传输耗时，即当前时间减去父块区块头中的时间戳
 	passTime := currentTime - parentTime
 	// 可以出块的时间窗口
-	windowFrom, windowTo := consensus.GetNextMineWindow(mineHeight, distance, parentTime, currentTime, m.timeoutTime, m.dm)
+	windowFrom, windowTo := m.dm.GetNextMineWindow(mineHeight, distance, parentTime, currentTime, m.timeoutTime)
 
 	if distance == 1 && passTime < m.timeoutTime {
 		// distance == 1表示下一个区块该本节点产生了，也没有超时，windowFrom为0。这时需要确保延迟足够的时间，避免早期交易少时链上全是空块
@@ -209,7 +208,7 @@ func (m *Miner) schedule(parentBlock *types.Block) bool {
 		return false
 	}
 	// 获取新块离本节点索引的距离，永远在(0,DeputyCount]区间中
-	distance, err := consensus.GetMinerDistance(mineHeight, parentBlock.MinerAddress(), minerAddress, m.dm)
+	distance, err := m.dm.GetMinerDistance(mineHeight, parentBlock.MinerAddress(), minerAddress)
 	if err != nil {
 		log.Errorf("GetMinerDistance error: %v", err)
 		return false
